backend/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.18
aiosignal==1.3.2
annotated-types==0.7.0
anthropic==0.49.0
anyio==4.9.0
attrs==25.3.0
backoff==2.2.1
beautifulsoup4==4.13.4
black==25.1.0
boto3==1.37.37
botocore==1.37.37
browser-use @ file:///Users/cpondoc/Desktop/classes/coterm/spring/cs349d/final-project/backend/browser-use
cachetools==5.5.2
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
Cython==3.0.12
defusedxml==0.7.1
deprecation==2.1.0
distro==1.9.0
faiss-cpu==1.10.0
fastapi==0.115.12
filetype==1.2.0
frozenlist==1.6.0
google-ai-generativelanguage==0.6.17
google-api-core==2.24.2
google-auth==2.39.0
googleapis-common-protos==1.70.0
gotrue==2.12.0
greenlet==3.2.0
grpcio==1.72.0rc1
grpcio-status==1.72.0rc1
h11==0.14.0
h2==4.2.0
hpack==4.1.0
httpcore==1.0.8
httpx==0.28.1
hyperframe==6.1.0
idna==3.10
iniconfig==2.1.0
jiter==0.9.0
jmespath==1.0.1
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.3.22
langchain-anthropic==0.3.3
langchain-aws==0.2.19
langchain-core==0.3.49
langchain-deepseek==0.1.3
langchain-google-genai==2.1.2
langchain-ollama==0.3.0
langchain-openai==0.3.11
langchain-text-splitters==0.3.7
langsmith==0.3.32
markdownify==1.1.0
mem0ai==0.1.93
monotonic==1.6
multidict==6.4.3
mypy_extensions==1.1.0
numpy==1.26.4
ollama==0.4.8
openai==1.75.0
orjson==3.10.16
packaging==24.2
patchright==1.51.3
pathspec==0.12.1
pillow==11.2.1
platformdirs==4.3.7
playwright==1.51.0
pluggy==1.5.0
portalocker==2.10.1
postgrest==1.0.1
posthog==3.25.0
propcache==0.3.1
proto-plus==1.26.1
protobuf==6.30.2
psutil==7.0.0
psycopg2-binary==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.10.6
pydantic_core==2.27.2
pyee==12.1.1
PyJWT==2.10.1
pyobjc-core==11.0
pyobjc-framework-Cocoa==11.0
pyperclip==1.9.0
pytest==8.3.5
pytest-mock==3.14.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2024.2
PyYAML==6.0.2
qdrant-client==1.14.2
realtime==2.4.3
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
rsa==4.9.1
s3transfer==0.11.5
screeninfo==0.8.1
six==1.17.0
sniffio==1.3.1
soupsieve==2.7
SQLAlchemy==2.0.40
starlette==0.46.2
storage3==0.11.3
StrEnum==0.4.15
supabase==2.15.1
supafunc==0.9.4
tenacity==9.1.2
tiktoken==0.9.0
tqdm==4.67.1
typing_extensions==4.13.2
urllib3==1.26.20
uvicorn==0.34.2
websockets==14.2
yarl==1.20.0
zstandard==0.23.0
